pyproject.toml
[project]
name = "platform-questions-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.10.1",
    "dj-database-url>=3.1.0",
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-stubs>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.14",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
]


[tool.black]
line_length = 90

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 90

[tool.ruff.lint]
select = ["E", "F", "W", "ANN", "RUF"]
ignore = ["D100", "D101"]

[tool.ruff.lint.per-file-ignores]
"config/settings/*.py" = ["F403", "F405"]
"apps/**/migrations/*.py" = ["RUF012", "E501"]





[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
ignore_missing_imports = true
strict = true
follow_imports = "silent"
disable_error_code = ["untyped-decorator"]

[tool.django-stubs]
django_settings_module = "config.settings.dev"
strict_settings = true
strict_model_abstract_attrs = true


[[tool.mypy.overrides]]
module = "rest_framework.serializers"
ignore_errors = true


[dependency-groups]
dev = [
    "black>=25.12.0",
    "django-stubs>=5.2.8",
    "djangorestframework-stubs>=3.16.7",
    "mypy>=1.19.1",
    "ruff>=0.14.13",
]
